backend/requirements.txt
flask
python-dotenv
google-generativeai
langchain
faiss-cpu
pandas